+++ src/components/ConnectionListing/ConnectionListing.js
@@ -7,16 +7,21 @@ import {
   Layout,
   MultiColumnList,
 } from '@folio/stripes/components';
-import { AmountWithCurrencyField } from '@folio/stripes-acq-components';
+import {
+  acqRowFormatter,
+  AmountWithCurrencyField,
+} from '@folio/stripes-acq-components';
 
-const defaultColumns = ['name', 'code', 'allocated', 'unavailable', 'available'];
+const defaultColumns = ['name', 'code', 'allocated', 'unavailable', 'available', 'arrow'];
 const columnMapping = {
+  arrow: null,
   name: <FormattedMessage id="ui-finance.item.name" />,
   code: <FormattedMessage id="ui-finance.item.code" />,
   allocated: <FormattedMessage id="ui-finance.item.allocated" />,
   unavailable: <FormattedMessage id="ui-finance.item.unavailable" />,
   available: <FormattedMessage id="ui-finance.item.available" />,
 };
+const alignRowProps = { alignLastColToEnd: true };
 
 const ConnectionListing = ({ items, currency, openItem, visibleColumns }) => {
   const resultsFormatter = {
@@ -38,6 +43,7 @@ const ConnectionListing = ({ items, currency, openItem, visibleColumns }) => {
         currency={currency}
       />
     ),
+    arrow: () => <Icon icon="caret-right" />,
   };
 
   return (
@@ -47,6 +53,8 @@ const ConnectionListing = ({ items, currency, openItem, visibleColumns }) => {
         contentData={items}
         formatter={resultsFormatter}
         onRowClick={openItem}
+        rowFormatter={acqRowFormatter}
+        rowProps={alignRowProps}
         visibleColumns={visibleColumns}
       />
       <Layout className="textCentered">